[vim] add stay-centered plugin, rename conf files
