+++ vim/.config/nvim/lua/plugs/stay-centered.lua
@@ -1,0 +1,1 @@
+require("stay-centered")
